Change mmo field to numeric

+++ database/001_schema.sql
@@ -98,7 +98,7 @@ create table if not exists tracks (
     jam_code text,
 
     -- Misc Epic fields
-    mmo integer,
+    mmo numeric,
     ci numeric,
     ag text,
     sm text,
